Update Tmux Config
Adjusts the Tmux config to no longer use TPM, instead relying on the default functionality availabile within Tmux itself. This prevents the need for Nerd Fonts to be present for Powerline rendering and leaves the basic status information up to Neovim where we actually care about that information being presented.